separated toggle for advanced options and the debug feed. Deleted ConfigMenuElement and restructured ConfigMenu. Rewrote tests to match new structure of ConfigMenu and its children
