+++ src/__tests__/components/ConfigMenu.test.tsx
@@ -1,19 +1,27 @@
 import { mount, shallow } from 'enzyme';
 import jsdom from 'jsdom';
 import React from 'react';
+import { Provider } from 'react-redux';
+import { Store } from 'redux';
 import ConfigMenu, {
+    ConfigMenuProps,
     IConfigMenuProps,
 } from '../../components/configMenu/ConfigMenu';
 import { HelpWith } from '../../components/configMenu/Help';
 import NumberMenuItem, {
     INumberMenuItemProps,
 } from '../../components/configMenu/menuItems/NumberMenuItem';
+import { initialState } from '../../store/reducers/configReducer';
+import { createStore } from '../../store/store';
 
 let props: IConfigMenuProps;
 let numberProps: INumberMenuItemProps;
 let children: React.ReactNode;
 let window: Window;
+import configureStore from 'redux-mock-store';
+import thunk from 'redux-thunk';
 let map: { [key: string]: any };
+let store: Store;
 
 describe('ConfigMenu', () => {
     beforeEach(() => {
@@ -34,35 +42,51 @@ describe('ConfigMenu', () => {
             map[event] = cb;
         });
         props = {
-            width: '15',
-            timerLength: 1,
-            children,
             window,
-            debugEnabled: false,
         };
+        const mockStore = configureStore([thunk]);
+        store = mockStore(initialState);
     });
 
     it('should render correctly', () => {
-        const wrapper = shallow(<ConfigMenu {...props} />).debug();
+        const wrapper = shallow(
+            <Provider store={store}>
+                <ConfigMenu {...props} />
+            </Provider>,
+        ).debug();
         expect(wrapper).toMatchSnapshot();
     });
 
     it('should set mousemove listener on mount', () => {
-        mount(<ConfigMenu {...props} />);
+        mount(
+            <Provider store={store}>
+                <ConfigMenu {...props} />
+            </Provider>,
+        );
         expect(map).toHaveProperty('mousemove');
     });
 
     it('should call mouseMoveHandler', () => {
-        const wrapper = mount(<ConfigMenu {...props} />);
-        wrapper.simulate('mousemove');
-        expect(wrapper.state('isUnderMouse')).toBe(false);
+        const wrapper = mount(
+            <Provider store={store}>
+                <ConfigMenu {...props} />
+            </Provider>,
+        );
+        const child = wrapper.find('ConfigMenu').first();
+        child.simulate('mousemove');
+        expect(child.state('isUnderMouse')).toBe(false);
     });
 
     it('should call onMouseLeave', () => {
-        const wrapper = mount(<ConfigMenu {...props} />);
+        const wrapper = mount(
+            <Provider store={store}>
+                <ConfigMenu {...props} />
+            </Provider>,
+        );
+        const child = wrapper.find('ConfigMenu').first();
         wrapper.simulate('mouseenter');
-        expect(wrapper.state('isUnderMouse')).toBe(true);
+        expect(child.state('isUnderMouse')).toBe(true);
         wrapper.simulate('mouseleave');
-        expect(wrapper.state('isUnderMouse')).toBe(false);
+        expect(child.state('isUnderMouse')).toBe(false);
     });
 });
