+++ src/store/actions/config/types.ts
@@ -27,6 +27,7 @@ export interface IAppConfig {
     ySensitivity: number;
     fps: number;
     toggleDebug: boolean;
+    toggleAdvanced: boolean;
     irisColor: string;
     toggleReflection: boolean;
     reflectionOpacity: number;
