+++ src/store/reducers/configReducer.ts
@@ -33,6 +33,7 @@ export const initialState: IConfigState = {
     detectionConfig: initialDetectionConfig,
     toggleReflection: true,
     reflectionOpacity: 0.2,
+    toggleAdvanced: false,
 };
 
 const configActionMapping = {
